backend/requirements.txt
chromadb
torch
transformers
pillow
uagents
google-generativeai
opencv-python
flask
flask_cors
